requirements_server.txt
python-socketio==5.11.0
uvicorn==0.27.0